feat: move `type_check_binary_operation` and `check_dtypes` to `util.rs`
