+++ crates/proof-of-sql/src/sql/mod.rs
@@ -11,3 +11,4 @@ pub mod proof;
 pub mod proof_exprs;
 pub mod proof_gadgets;
 pub mod proof_plans;
+pub(crate) mod util;
